requirements.txt
discord.py
discord.py[voice]
python-dotenv
youtube-dl
google-api-python-client